requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crontab==0.7.1
django-kronos==1.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
kronos==0.2.2
logger==1.4
numpy==1.20.2
pandas==1.2.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
